requirements_dev.txt
-r requirements.txt

black>=22.10.0,<22.11
isort>=5.10.1,<5.11
flake8>=4.0.1,<4.1
wemake-python-styleguide>=0.17.0,<0.18
mypy==0.982
djangorestframework-stubs>=1.7.0,<1.8
pylint-django>=2.5.3,<2.6
pytest-django>=4.5.2,<4.6
coverage>=6.5.0,<6.6
